权限管理

+++ src/main/java/com/breech/extremity/mapper/RoleMapper.java
@@ -9,6 +9,6 @@ import java.util.List;
 public interface RoleMapper extends Mapper<Role> {
     List<Role> selectRoleByIdUser(@Param("id") Long id);
     Role selectRoleByInputCode(@Param("inputCode") String inputCode);
-    Integer updateStatus(@Param("idRole") Long idRole, @Param("status") String status);
-    Integer update(@Param("idRole") Long idRole, @Param("name") String name, @Param("inputCode") String inputCode, @Param("weights") Integer weights);
+    Integer updateStatus(@Param("idRole") Integer idRole, @Param("status") String status);
+    Integer update(@Param("idRole") Integer idRole, @Param("name") String name, @Param("inputCode") String inputCode, @Param("weights") Integer weights);
 }
